move tests of ingest into it's own file
remove unused variables

add to_dict method for animal_sets

create partial for changeset template

return the stringified version of the foreign key of an object

make valid tables w/ 1 thead/table
