+++ zootable/zoo_checks/templates/tally.html
@@ -11,9 +11,8 @@
     {{animals_formset.management_form}}
     {{groups_formset.management_form}}
 
-    <table class="striped">
     {% for spec_dict in formset_order.values %}
-        <tbody>
+    <table class="striped">
         {% include "species_count_form_snippet.html" with form=spec_dict.formset species=spec_dict.species prior_counts=spec_dict.prior_counts %}
 
         {% if spec_dict.group_form %}
@@ -27,9 +26,8 @@
             {% include "animal_form_snippet.html" with form=ff %}
             </tr>
         {% endfor %}
-        </tbody>
-    {% endfor %}
     </table>
+    {% endfor %}
 
     <div class="fixed-action-btn">
         <button class="btn-floating btn-large waves-effect waves-light red" type="submit" name="action">
